Introduced individual key redefinition dialogues.
Instead of the ad-hoc dialogue showing a message and reading the
keyboard, a simple dialogue is used, showing defined keys in a
listbox and providing buttons to assign and remove definitions,
also providing "OK" and "Cancel" buttons. Thus, individual key
changes can be made or discarded as can whole keyboard changes.