Support parallel output via a socket.
As an alternative to sending parallel port data to standard output,
it may be directed to a socket for further processing, as is
already done with serial port data. The tools/serial_client.py
program can be used to monitor parallel port output in the same
fashion as serial port output.