Fixed interrupt handling with Plus 1 disabled.
Since the "master IRQ" bit in the interrupt status register is set
in the emulator when other conditions occur, it cannot normally be
used by itself as the basis for initiating interrupt handling. If
this bit is tested for such purposes, an emulated unexpanded
machine will not start up successfully because nothing will clear
this bit once it has been set.

However, a mechanism is still needed for peripherals to signal
interrupt conditions. Here, the workaround is to test for the Plus
1 which will let the serial polling function disable the bit when
the serial controller is not signalling an interrupt condition,
with the Plus 1 software being able to cancel any such condition.