Use conventional exit statuses
The code uses "exit(-1)" to indicate an error in various places. It's
normal to "exit(1)" (i.e. "exit(EXIT_FAILURE)") on error, and this is
less messy if, e.g., you've got a shell prompt that shows you the exit
status of the previous command.

For "--help", the convention is to exit(0), since the program has
successfully done what you've asked for.